requirements.txt
# Installation de l'environement virtuel avec 'python3 -m venv --system-site-packages venv'
# Raspberry 5
numpy>=2.2.5
pygame>=2.6.1
sounddevice>=0.5.1
pydub==0.25.1
pyalsaaudio==0.11.0